OPTIMIZE: move kline.go into x/stream/common/kline
